Delete Passenger Api. Yet to test due to non complication of reservation api's

+++ src/main/java/com/cmpe275/AirlineReservationSystem/service/PassengerService.java
@@ -1,12 +1,17 @@
 package com.cmpe275.AirlineReservationSystem.service;
 
+import com.cmpe275.AirlineReservationSystem.entity.Flight;
 import com.cmpe275.AirlineReservationSystem.entity.Passenger;
+import com.cmpe275.AirlineReservationSystem.entity.Reservation;
 import com.cmpe275.AirlineReservationSystem.repository.PassengerRepository;
+import com.cmpe275.AirlineReservationSystem.repository.ReservationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import java.util.*;
+
 
 import java.util.Optional;
 
@@ -14,6 +19,9 @@ import java.util.Optional;
 public class PassengerService {
     @Autowired
     private PassengerRepository passengerRepository;
+
+    @Autowired
+    private ReservationRepository reservationRepository;
 
     public Passenger createPassenger(Passenger passenger){
         return passengerRepository.save(passenger);
@@ -41,8 +49,40 @@ public class PassengerService {
         //return new ResponseEntity<>(, HttpStatus.OK);
     }
 
-    public boolean deletePassenger(String id){
-         passengerRepository.deleteById(id);
-         return true;
+    public void deleteReservation(Reservation reservation, Passenger passenger){
+        try{
+            for(Flight flight : reservation.getFlights()){
+                updateFlightSeats(flight);
+                flight.getPassengers().remove(passenger);
+            }
+            passenger.getReservation().remove(reservation);
+            reservationRepository.delete(reservation);
+        }
+        catch(Exception e){
+            System.out.println("Exception");
+        }
+    }
+
+    public void updateFlightSeats(Flight flight){
+        try{
+            flight.setSeatsLeft(flight.getSeatsLeft()+1);
+        }
+        catch(Exception e){
+
+        }
+    }
+
+    public ResponseEntity<?> deletePassenger(String id){
+         Optional<Passenger> existingPass=passengerRepository.findById(id);
+        if(existingPass.isPresent()){
+            List<Reservation> reservations = reservationRepository.findByPassengerID(existingPass.get());
+            for(Reservation reservation : reservations){
+                deleteReservation(reservation, existingPass.get());
+            }
+            passengerRepository.deleteById(id);
+            return new ResponseEntity<>("Passenger with id"+ id+ " is deleted successfully ", HttpStatus.OK);
+        }else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Passenger Does not exist");
+        }
     }
 }
